refactor(drawer): move context into an internal directory
Moves `DrawerContext` into an `__internal__` directory and updates imports where neccessary. The
context is not intended for public use and should improve tree-shaking in Carbon.

BREAKING CHANGE: The context is no longer exposed as part of the public interface of `Drawer`
