+++ src/components/drawer/__internal__/drawer-sidebar.context.ts
@@ -1,0 +1,9 @@
+import React from "react";
+
+export interface DrawerSidebarContextProps {
+  isInSidebar: boolean;
+}
+
+export default React.createContext<DrawerSidebarContextProps>({
+  isInSidebar: false,
+});
