+++ src/components/drawer/index.ts
@@ -1,5 +1,2 @@
-export { default, DrawerSidebarContext } from "./drawer.component";
-export type {
-  DrawerSidebarContextProps,
-  DrawerProps,
-} from "./drawer.component";
+export { default } from "./drawer.component";
+export type { DrawerProps } from "./drawer.component";
