+++ src/components/tabs/tabs.spec.tsx
@@ -15,7 +15,7 @@ import {
   testStyledSystemMargin,
 } from "../../__spec_helper__/__internal__/test-utils";
 import { StyledTabsHeaderWrapper } from "./__internal__/tabs-header/tabs-header.style";
-import { DrawerSidebarContext } from "../drawer";
+import DrawerSidebarContext from "../drawer/__internal__/drawer-sidebar.context";
 import Textbox from "../textbox";
 import NumeralDate from "../numeral-date";
 import ValidationIcon, {
